Device Parameters

DAQ

MPIMeasurements.RedPitayaDAQParams — Type

Parameters for a DAQ of type RedPitayaDAQ

channels: All configured channels of this DAQ device.

ips: IPs of the Red Pitayas

triggerMode: Trigger mode of the Red Pitayas. Default: ALL_INTERNAL.

resetWaittime: Time to wait after a reset has been issued.

rampingMode

rampingFraction

useCounterTrigger: Flag for using the counter trigger

counterTriggerSourceType: Source type of the counter trigger

counterTriggerSourceChannel: DIO pin used for the counter trigger

MPIMeasurements.RedPitayaLUTChannelParams — Type

Parameters for a TxChannel of type RedPitayaLUTChannel

channelIdx

calibration

range

hbridge

switchTime

switchEnable

MPIMeasurements.DAQRxChannelParams — Type

Parameters for a general DAQRxChannel

channelIdx: required, Integer, channel idx of the corresponding DAQ

transferFunction: optional, can be either a filename to a .h5 file in the TransferFunctions folder or a single (complex) number

MPIMeasurements.DAQTxChannelParams — Type

Parameters for a general DAQTxChannel

channelIdx

limitPeak

limitSlewRate

sinkImpedance

allowedWaveforms

feedbackChannelID

calibration

TxController

MPIMeasurements.TxDAQControllerParams — Type

Parameters for a TxDAQController`

phaseAccuracy: Angle, required, allowed deviation of the excitation phase

relativeAmplitudeAccuracy: Number, required, allowed relative deviation of the excitation amplitude

absoluteAmplitudeAccuracy: Magnetic field, default: 50µT, allowed absolute deviation of the excitation amplitude

maxControlSteps: Integer, default: 20, maximum number of steps to try to control the system

controlDC: Bool, default: false, control the DC value of the excitation field (only posible for DC enabled DF amplifiers)

timeUntilStable: Float, default: 0.0, time in seconds to wait before the DF is stable after ramping

minimumStepDuration: Float, default: 0.002, time in seconds that the DF should be averaged during the control measurement

fieldToVoltRelDeviation: Float, default: 0.2, relative deviation allowed between forward calibration and actual system state

fieldToVoltAbsDeviation: Magnetic field, default: 5.0mT, absolute deviation allowed between forward calibration and actual system state

maxField: Magnetic field, default: 40mT, maximum field amplitude that the controller should allow