Protocols

MPIMeasurements.ContinousMeasurementProtocolParams — Type

Parameters for the ContinousMeasurementProtocol`

fgFrames: Foreground frames to measure. Overwrites sequence frames

bgFrames: Background frames to measure. Overwrites sequence frames

pause: Pause between measurements

controlTx: If set the tx amplitude and phase will be set with control steps

measureBackground: If unset no background measurement will be taken

sequence: Sequence to measure

MPIMeasurements.MPIForceProtocolParams — Type

Parameters for the MPIForceProtocol

controlTx: If set the tx amplitude and phase will be set with control steps

sequence: Sequence to measure

MPIMeasurements.MPIMeasurementProtocolParams — Type

Parameters for the MPIMeasurementProtocol

fgFrames: Foreground frames to measure. Overwrites sequence frames

bgFrames: Background frames to measure. Overwrites sequence frames

controlTx: If set the tx amplitude and phase will be set with control steps

measureBackground: If unset no background measurement will be taken

saveTemperatureData: If the temperature should be safed or not

sequence: Sequence to measure

reuseLastBGMeas: Do not measure background but reuse the last BG measurement if suitable

MPIMeasurements.MPSMeasurementProtocolParams — Type

Parameters for the MPSMeasurementProtocol

fgFrames: Foreground frames to measure. Overwrites sequence frames

bgFrames: Background frames to measure. Overwrites sequence frames

controlTx: If set the tx amplitude and phase will be set with control steps

measureBackground: If unset no background measurement will be taken

rememberBGMeas: Remember background measurement

saveTemperatureData: If the temperature should be saved or not

sequence: Sequence to measure

sortPatches: Sort patches

saveInCalibFolder: Flag if the measurement should be saved as a system matrix or not

dfPeriodsPerOffset: Number of periods per offset of the MPS offset measurement. Overwrites parts of the sequence definition.

averagePeriodsPerOffset: If true all periods per offset are averaged

MPIMeasurements.MechanicalMPIMeasurementProtocolParams — Type

Parameters for the MechanicalMPIMeasurementProtocol

fgFrames: Foreground frames to measure. Overwrites sequence frames

bgFrames: Background frames to measure. Overwrites sequence frames

controlTx: If set the tx amplitude and phase will be set with control steps

measureBackground: If unset no background measurement will be taken

saveTemperatureData: If the temperature should be safed or not

sequence: Sequence to measure

rememberBGMeas: Remember background measurement

MPIMeasurements.MultiSequenceSystemMatrixProtocolParams — Type

Parameters for the MultiSequenceSystemMatrixProtocol

controlTx: If set the tx amplitude and phase will be set with control steps

saveTemperatureData: If the temperature should be safed or not

sequences: Sequences to measure

positions: SM Positions mapped to the natural sorting of sequence tomls

saveInCalibFolder: Flag if the calibration should be saved as a system matrix or not

waitTime: Seconds to wait between measurements

coolDownTo: Maximum temperature the coils can be at to start the next measurement

bgSequence: Sequence to use as a BG measurement

numBGMeas: Number of BG measurements before and after the measurement

MPIMeasurements.RobotBasedMagneticFieldStaticProtocolParams — Type

Parameter for the RobotBasedMagneticFieldStaticProtocol

sequence

positions

description

MPIMeasurements.RobotBasedSystemMatrixProtocolParams — Type

Parameters for the RobotBasedSystemMatrixProtocol

waitTime: Minimum wait time between robot movements

bgFrames: Number of background frames to measure for a background position

fgFrames: Number of frames that are averaged for a foreground position

saveAsSystemMatrix: Flag if the calibration should be saved as a system matrix or not

saveTemperatureData: Flag if the temperature measured after every robot measurement should be stored in the MDF or not

bgMeas: Number of background measurements to take

controlTx: If set the tx amplitude and phase will be set with control steps

sequence: Sequence used for the calibration at each position

positions

MPIMeasurements.RobotBasedTDesignFieldProtocolParams — Type

Parameters for the RobotBasedTDesignFieldProtocol

sequence

radius

center

N

T

MPIMeasurements.RobotMPIMeasurementProtocolParams — Type

Parameters for the RobotMPIMeasurementProtocol

fgPos: Foreground position

fgFrames: Foreground frames to measure. Overwrites sequence frames

bgFrames: Background frames to measure. Overwrites sequence frames

controlTx: If set the tx amplitude and phase will be set with control steps

measureBackground: If unset no background measurement will be taken

saveTemperatureData: If the temperature should be safed or not

sequence: Sequence to measure

rememberBGMeas: Remember background measurement