Sequence Parameters

Magnetic Field

MPIMeasurements.MagneticField — Type

Description of a magnetic field.

The field can either be electromagnetically or mechanically changed. The mechanical movement of e.g. an iron yoke would be defined within two channels, one electrical and one mechanical.

id: Unique ID of the field description.

channels: Transmit channels that are used for the field.

safeStartInterval: Flag if the start of the field should be convoluted. If the DAQ does not support this, it can may fall back to postponing the application of the settings. Not used for mechanical fields.

safeTransitionInterval: Flag if a transition of the field should be convoluted. If the DAQ does not support this, it can may fall back to postponing the application of the settings. Not used for mechanical fields.

safeEndInterval: Flag if the end of the field should be convoluted. In case of an existing brake on a mechanical channel this means a use of the brake.

safeErrorInterval: Flag if the field should be convoluted down in case of an error. In case of an existing brake on a mechanical channel this means a use of the brake.

control: Flag if the channels of the field should be controlled.

decouple: Flag if the field should be decoupled. Not used for mechanical channels.

Channels

MPIMeasurements.ContinuousElectricalChannel — Type

Electrical channel with a stepwise definition of values.

MPIMeasurements.ContinuousMechanicalRotationChannel — Type

Mechanical channel with a continuous rotation.

MPIMeasurements.ContinuousMechanicalTranslationChannel — Type

Mechanical channel describing a continuous translational movement.

MPIMeasurements.PeriodicElectricalChannel — Type

Electrical channel based on based on periodic base functions. Only the PeriodicElectricalChannel counts for the cycle length calculation

id: ID corresponding to the channel configured in the scanner.

components: Components added for this channel.

offset: Offset of the channel. If defined in Tesla, the calibration configured in the scanner will be used.

isDfChannel

dcEnabled

MPIMeasurements.StepwiseElectricalChannel — Type

Electrical channel with a stepwise definition of values.

id: ID corresponding to the channel configured in the scanner.

divider: Divider of the component.

values: Values corresponding to the individual steps.

enable: TBD

MPIMeasurements.StepwiseMechanicalRotationChannel — Type

Mechanical channel with a triggered stepwise rotation.

MPIMeasurements.StepwiseMechanicalTranslationChannel — Type

Mechanical channel describing a stepwise translational movement.

Components

MPIMeasurements.ArbitraryElectricalComponent — Type

Arbitrary waveform component of an electrical channel with periodic base function defined by a vector of 16384 values.

id

divider: Divider of the component.

amplitude: Amplitude scale of the base waveform for each period of the field

phase: Phase of the component for each period of the field.

values: Values for the base waveform of the component, will be multiplied by amplitude

MPIMeasurements.PeriodicElectricalComponent — Type

Component of an electrical channel with periodic base function.

id

divider: Divider of the component.

amplitude: Amplitude (peak) of the component for each period of the field.

phase: Phase of the component for each period of the field.

waveform: Waveform of the component.

MPIMeasurements.SweepElectricalComponent — Type

Sweepable component of an electrical channel with periodic base function. Note: Does not allow for changes in phase since this would make the switch between frequencies difficult.

divider: Divider of the component.

amplitude: Amplitude (peak) of the channel for each divider in the sweep. Must have the same dimension as divider.

waveform: Waveform of the component.