Scanner Construction

MPIMeasurements.MPIScanner — Type

Basic description of a scanner.

name: Name of the scanner

configFile: Path to the used configuration file.

generalParams: General parameters of the scanner like its bore size or gradient.

devices: Device instances instantiated by the scanner from its configuration.

MPIMeasurements.name — Function

Name of the scanner

MPIMeasurements.configDir — Function

Path to the used configuration directory.

Retrieve the configuration directory of the scanner that the device is contructed from.

MPIMeasurements.generalParams — Function

General parameters of the scanner like its bore size or gradient.

MPIMeasurements.scannerBoreSize — Function

Bore size of the scanner.

MPIMeasurements.scannerFacility — Function

Facility where the scanner is located.

MPIMeasurements.scannerManufacturer — Function

Manufacturer of the scanner.

MPIMeasurements.scannerName — Function

Name of the scanner

MPIMeasurements.scannerTopology — Function

Topology of the scanner, e.g. FFL or FFP.

MPIMeasurements.scannerGradient — Function

Gradient of the scanners selection field.

MPIMeasurements.scannerDatasetStore — Function

Path of the dataset store.

MPIMeasurements.defaultProtocol — Function

Default protocol of the scanner.

Device Handling

MPIMeasurements.getDevices — Method

getDevices(scanner, deviceType)

Retrieve all devices of a specific deviceType. Returns an empty vector if none are found

Missing docstring.

Missing docstring for MPIMeasurements.getDevices(::MPIScanner, ::String). Check Documenter's build log for details.

MPIMeasurements.getDevice — Method

getDevice(scanner, deviceType)

Retrieve a device of a specific deviceType if it can be unambiguously retrieved. Returns nothing if no such device can be found and throws an error if multiple devices fit the type.

MPIMeasurements.getDeviceIDs — Function

getDeviceIDs(scanner)

Retrieve a list of all device IDs available for the scanner.

Sequence Handling

MPIMeasurements.Sequence — Method

Sequence(scanner, name)

Constructor for a sequence of name from the configuration directory specified for the scanner.

MPIMeasurements.getSequenceList — Function

getSequenceList(scanner)

Retrieve a list of all sequences available for the scanner.

Protocol Handling

Missing docstring.

Missing docstring for MPIMeasurements.Protocol(::AbstractString, ::MPIScanner). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPIMeasurements.getProtocolList. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPIMeasurements.execute(::MPIScanner, ::Protocol). Check Documenter's build log for details.