Public Documentation

Documentation for MPIMeasurements.jl's public interface.

See the Internals section of the manual for internal package docs covering all submodules.

Contents

Public Documentation

Contents

Index

Public Interface

Index

MPIMeasurements.concreteTypesCache

MPIMeasurements.iselErrorCodes

MPIMeasurements.AbstractDeviceException

MPIMeasurements.AcquisitionSettings

MPIMeasurements.BrukerCommand

MPIMeasurements.ConfigurationError

MPIMeasurements.Device

MPIMeasurements.Device

MPIMeasurements.DeviceException

MPIMeasurements.DeviceParams

MPIMeasurements.FOTemp

MPIMeasurements.IgusRobotParams

MPIMeasurements.IselRobotPoolParams

MPIMeasurements.IselRobotPortParams

MPIMeasurements.LakeShoreF71GaussMeter

MPIMeasurements.Protocol

MPIMeasurements.Protocol

MPIMeasurements.ProtocolConfigurationError

MPIMeasurements.ProtocolOffsetElectricalChannel

MPIMeasurements.RedPitayaDAQParams

MPIMeasurements.RxChannel

MPIMeasurements.ScannerConfigurationError

MPIMeasurements.Sequence

MPIMeasurements.Sequence

MPIMeasurements.SequenceConfigurationError

MPIMeasurements.SimpleBoreCollisionModuleParams

MPIMeasurements.SimpleSimulatedDAQParams

MPIMeasurements.TCPSCPIInstrument

MPIMeasurements.Waveform

Base.close

Base.close

Base.show

Base.show

MPIMeasurements.Devices

MPIMeasurements.__init__

MPIMeasurements._moveRel

MPIMeasurements.checkCollisionYZCircle

MPIMeasurements.checkCoords

MPIMeasurements.checkCoords

MPIMeasurements.checkPosition

MPIMeasurements.command

MPIMeasurements.command

MPIMeasurements.configDir

MPIMeasurements.configDir

MPIMeasurements.configFile

MPIMeasurements.configFile

MPIMeasurements.createDAQChannels

MPIMeasurements.createDAQParams

MPIMeasurements.createModbusTelegram

MPIMeasurements.createModbusTelegram

MPIMeasurements.createRFFTindices

MPIMeasurements.createReferenceIndexMapping

MPIMeasurements.deepsubtypes

MPIMeasurements.defaultProtocol

MPIMeasurements.deviceID

MPIMeasurements.expectedDependencies

MPIMeasurements.generalParams

MPIMeasurements.getActiveChannel

MPIMeasurements.getActiveChannels

MPIMeasurements.getAllFields

MPIMeasurements.getAveragedTemperature

MPIMeasurements.getAveragedTemperature

MPIMeasurements.getControlResult

MPIMeasurements.getDevice

MPIMeasurements.getDevice

MPIMeasurements.getDeviceIDs

MPIMeasurements.getDevices

MPIMeasurements.getFast

MPIMeasurements.getFirmwareVersion

MPIMeasurements.getMeasuringChannels

MPIMeasurements.getMode

MPIMeasurements.getModelName

MPIMeasurements.getPeakRMS

MPIMeasurements.getProbeType

MPIMeasurements.getRange

MPIMeasurements.getSdoObject

MPIMeasurements.getSequenceList

MPIMeasurements.getSerialDevices

MPIMeasurements.getSerialNumber

MPIMeasurements.getTemperature

MPIMeasurements.getTemperature

MPIMeasurements.getTemperatureAveraging

MPIMeasurements.getUnit

MPIMeasurements.getXYZValues

MPIMeasurements.getXYZValues

MPIMeasurements.hasDependency

MPIMeasurements.identification

MPIMeasurements.init

MPIMeasurements.initAxes

MPIMeasurements.initiateDevices

MPIMeasurements.invertAxesYZ

MPIMeasurements.invertAxisZ

MPIMeasurements.isAutoRanging

MPIMeasurements.isExpectedDependency

MPIMeasurements.isFieldComp

MPIMeasurements.isInSleepMode

MPIMeasurements.isLocked

MPIMeasurements.isOptional

MPIMeasurements.isPresent

MPIMeasurements.isProbeOn

MPIMeasurements.isTempComp

MPIMeasurements.lockOff

MPIMeasurements.lockOn

MPIMeasurements.moveCenter

MPIMeasurements.movePark

MPIMeasurements.name

MPIMeasurements.numChannels

MPIMeasurements.params

MPIMeasurements.query

MPIMeasurements.query

MPIMeasurements.query

MPIMeasurements.query

MPIMeasurements.query

MPIMeasurements.queryIsel

MPIMeasurements.readIOInput

MPIMeasurements.readIOInput

MPIMeasurements.readModbusTelegram

MPIMeasurements.receive

MPIMeasurements.scannerBoreSize

MPIMeasurements.scannerDatasetStore

MPIMeasurements.scannerFacility

MPIMeasurements.scannerGradient

MPIMeasurements.scannerManufacturer

MPIMeasurements.scannerName

MPIMeasurements.scannerTopology

MPIMeasurements.selfTest

MPIMeasurements.sendAndReceiveTelegram

MPIMeasurements.sendCommand

MPIMeasurements.setAcceleration

MPIMeasurements.setActiveChannel

MPIMeasurements.setActiveChannel

MPIMeasurements.setAllAutoRanging

MPIMeasurements.setAllMode

MPIMeasurements.setAllRange

MPIMeasurements.setAutoRanging

MPIMeasurements.setBrake

MPIMeasurements.setCompleteProbe

MPIMeasurements.setFast

MPIMeasurements.setFieldComp

MPIMeasurements.setFree

MPIMeasurements.setFrontPanelLock

MPIMeasurements.setMode

MPIMeasurements.setPeakRMS

MPIMeasurements.setProbe

MPIMeasurements.setRange

MPIMeasurements.setRefVelocity

MPIMeasurements.setSdoObject

MPIMeasurements.setSleepMode

MPIMeasurements.setStandardSettings

MPIMeasurements.setStartStopFreq

MPIMeasurements.setTempComp

MPIMeasurements.setTemperatureAveraging

MPIMeasurements.setTxParams

MPIMeasurements.setUnit

MPIMeasurements.setUnitToGauss

MPIMeasurements.setUnitToTesla

MPIMeasurements.setXRange

MPIMeasurements.setYRange

MPIMeasurements.setZRange

MPIMeasurements.setZeroPoint

MPIMeasurements.set_delim_read

MPIMeasurements.set_delim_write

MPIMeasurements.set_timeout_ms

MPIMeasurements.simRefZYX

MPIMeasurements.simulateLangevinInduced

MPIMeasurements.sleepModeOff

MPIMeasurements.sleepModeOn

MPIMeasurements.toCurrentMode

MPIMeasurements.toHighVoltageMode

MPIMeasurements.toLowVoltageMode

MPIMeasurements.toVoltageMode

MPIMeasurements.writeIOOutput

Sockets.send

Public Interface

MPIMeasurements.concreteTypesCache — Constant

Retrieve the concrete type of a given supertype corresponding to a given string.

MPIMeasurements.iselErrorCodes — Constant

Errorcodes Isel Robot

MPIMeasurements.AbstractDeviceException — Type

Abstract exception type for exceptions in devices

MPIMeasurements.AcquisitionSettings — Type

Settings for acquiring the sequence.

MPIMeasurements.BrukerCommand — Type

BrukerCommand(command::String)

MPIMeasurements.ConfigurationError — Type

Abstract error type for errors in configuration files

MPIMeasurements.Device — Type

Abstract type for all devices

Every device has to implement its own device struct which identifies it. A concrete implementation should contain e.g. the handle to device ressources or internal variables. The device struct must at least have the fields deviceID, params and dependencies and all other fields should have default values.

MPIMeasurements.Device — Method

Device(scannerName, deviceID; kwargs...)

Initialize and retrieve a device by its deviceID from a scanners configuration file. Also initializes the devices dependency tree according to the initialization order defined in the configuration file.

MPIMeasurements.DeviceException — Type

General exception type for exceptions in devices

MPIMeasurements.DeviceParams — Type

Abstract type for all device parameters

Every device must implement a parameter struct allowing for automatic instantiation from the configuration file.

MPIMeasurements.FOTemp — Method

fotemp(portAdress::AbstractString)

Initialize Fotemp fiber optical temperature sensor. For an overview over the high level API call methodswith(SerialDevice{FOTemp}).

Low level functions such as getModelName(ft) will return a function hash and a whitespace, e.g. "#40 " leading the actual answer, which would be the model name of the device in this case.

MPIMeasurements.IgusRobotParams — Type

Parameters for a Robot of type IgusRobot

defaultVelocity

axisRange

ip

port

keepSocketOpen

stepsPermm

feed

shaftRev

homVelSwitch

homVelZero

homAcc

movAcc

movDec

timeout

namedPositions

coordinateSystem

MPIMeasurements.IselRobotPoolParams — Type

Parameters for a Robot of type IselRobot that is connected via a serial port pool

axisRange

defaultVel

defaultRefVel

invertAxes

minMaxVel

minMaxAcc

minMaxFreq

stepsPermm

roundmmDigits

description

delim_read

delim_write

baudrate

ndatabits

parity

nstopbits

timeout_ms

namedPositions

referenceOrder

movementOrder

coordinateSystem

enableWaitTime

MPIMeasurements.IselRobotPortParams — Type

Parameters for a Robot of type IselRobot that is connected directly via a serial port

axisRange

defaultVel

defaultRefVel

invertAxes

minMaxVel

minMaxAcc

minMaxFreq

stepsPermm

roundmmDigits

serial_port

delim_read

delim_write

baudrate

ndatabits

parity

nstopbits

timeout_ms

namedPositions

referenceOrder

movementOrder

coordinateSystem

enableWaitTime

MPIMeasurements.LakeShoreF71GaussMeter — Type

MPIMeasurements.Protocol — Type

Abstract type for all protocols

Every protocol has to implement its own protocol struct which identifies it. A concrete implementation should contain e.g. the handle to the datastore or internal variables. The device struct must at least have the fields name, description, scanner and params and all other fields should have default values.

MPIMeasurements.Protocol — Method

General constructor for all concrete subtypes of Protocol.

MPIMeasurements.ProtocolConfigurationError — Type

Specialized error type for errors in protocol configuration files

MPIMeasurements.ProtocolOffsetElectricalChannel — Type

Offset channnel that gets translated into the correct representation by the MPSMeasurementProtocol

id: ID corresponding to the channel configured in the scanner.

offsetStart

offsetStop

numOffsets

MPIMeasurements.RedPitayaDAQParams — Method

Create the params struct from a dict. Typically called during scanner instantiation.

MPIMeasurements.RxChannel — Type

Receive channel reference that should be included in the acquisition.

MPIMeasurements.ScannerConfigurationError — Type

Specialized error type for errors in scanner configuration files

MPIMeasurements.Sequence — Type

Description of a sequence that can be run by a scanner.

The sequence can either be continuous or triggered. Triggered in this context means that the acquisition is done on a certain event, e.g. the move of a robot. The sweeping of frequencies or movement points can also be done in a triggered or continuous fashion.

MPIMeasurements.Sequence — Method

Sequence(configdir, name)

Constructor for a sequence of name from configDir.

MPIMeasurements.SequenceConfigurationError — Type

Specialized error type for errors in sequence configuration files

MPIMeasurements.SimpleBoreCollisionModuleParams — Type

Parameters for a SimpleBoreCollisionModule`

scannerDiameter: Diameter of scanner in the y-z plane

objGeometry: Geometry of the probe, centered at (0,0,0) in scanner coordinates

clearance

minMaxBoreAxis: Define the minimum and maximum points of movement into and out of the bore axis (x)

MPIMeasurements.SimpleSimulatedDAQParams — Method

Create the params struct from a dict. Typically called during scanner instantiation.

MPIMeasurements.TCPSCPIInstrument — Method

Constructor interpreting the given ip string as a IPv4 address

MPIMeasurements.Waveform — Type

Enum describing the existing waveforms.

Base.close — Method

Close the serial port of the serial device sd.

Base.close — Method

close(scanner)

Close the devices when closing the scanner.

Base.show — Method

Base.show(io::IO, ex::ConfigurationError)

Custom printing of ConfigurationError subtypes

Base.show — Method

Base.show(io::IO, ex::DeviceException)

Custom printing of DeviceException

MPIMeasurements.Devices — Method

Devices(scannerName, deviceIDs; kwargs...)

Initialize and retrieve devices by theirdeviceID from a scanners configuration file. Also initializes the devices dependency trees according to the initialization order defined in the configuration file..

MPIMeasurements.__init__ — Method

__init__()

Initialize configuration paths with the package and enable MPI REPL mode.

MPIMeasurements._moveRel — Method

Not Implemented

MPIMeasurements.checkCollisionYZCircle — Method

Returns true, 0mm, 0mm if given geometry does not collide with the scanner radius based on given position and clearance

MPIMeasurements.checkCoords — Method

Convenience function to check an array of nx3 position vectors at once

MPIMeasurements.checkCoords — Method

if the function is called on a vector of collision modules return the logic AND of all decisions

MPIMeasurements.checkPosition — Method

Returns true if given position is allowed, false otherwise

MPIMeasurements.command — Method

Send a command to the instrument

MPIMeasurements.command — Method

Send a command to the instrument

MPIMeasurements.configDir — Method

Retrieve the configuration directory of the scanner that the device is contructed from.

MPIMeasurements.configDir — Method

Path to the used configuration directory.

MPIMeasurements.configFile — Method

Retrieve the path to the configuration file of a device.

MPIMeasurements.configFile — Method

Path to the used configuration file

MPIMeasurements.createDAQChannels — Method

Create DAQ channel description from device dict part.

MPIMeasurements.createDAQParams — Method

Create the params struct from a dict. Typically called during scanner instantiation.

MPIMeasurements.createModbusTelegram — Method

createModbusTelegram(sdoObject::SDOObj, data::Integer)

Creates a TCP payload that will write data into the given sdoObject

MPIMeasurements.createModbusTelegram — Method

createModbusTelegram(sdoObject::SDOObj)

Creates a TCP payload that will be identified as a read-only modbus command by the motor controller

MPIMeasurements.createRFFTindices — Method

This function creates the correct indices into a channel-wise rfft of the reference channels for each channel in the controlledChannelsDict

MPIMeasurements.createReferenceIndexMapping — Method

Returns a vector of indices into the the DAQ data after the channels have already been selected by the daq.refChanIDs to select the correct reference channels in the order of the controlledChannels

MPIMeasurements.deepsubtypes — Method

deepsubtypes(type)

Recursively find all concrete types of the given type.

MPIMeasurements.defaultProtocol — Method

Default protocol of the scanner.

MPIMeasurements.deviceID — Method

Retrieve the ID of a device.

MPIMeasurements.expectedDependencies — Method

Retrieve all expected dependencies of a device.

MPIMeasurements.generalParams — Method

General parameters of the scanner like its bore size or gradient.

MPIMeasurements.getActiveChannel — Method

Returns the active channel X, Y, Z, V = Vector Magnitude Channel

MPIMeasurements.getActiveChannels — Method

Returns the the active channels. The active and inactive channels are the bits of the ASCII encoded hexadecimal bytes returned after the whitespace following the function hash. I.e. "#10 1E" represents that channels 2, 3, 4, and 5 are on and the remaining channels 1,6,7, and 8 are switched off.

MPIMeasurements.getAllFields — Method

Returns the field values of X, Y and Z. Throws an error if an overload was detected

MPIMeasurements.getAveragedTemperature — Method

Returns the averaged temperature of a channel.

MPIMeasurements.getAveragedTemperature — Method

Returns the averaged temperature of all channel.

MPIMeasurements.getControlResult — Method

Returns a Sequence that is merged from the control result and all uncontrolled field of the given ControlSequence

MPIMeasurements.getDevice — Method

getDevice(scanner, deviceID)

Retrieve a device by its deviceID.

MPIMeasurements.getDevice — Method

getDevice(scanner, deviceType)

Retrieve a device of a specific deviceType if it can be unambiguously retrieved. Returns nothing if no such device can be found and throws an error if multiple devices fit the type.

MPIMeasurements.getDeviceIDs — Method

getDeviceIDs(scanner)

Retrieve a list of all device IDs available for the scanner.

MPIMeasurements.getDevices — Method

getDevices(scanner, deviceType)

Retrieve all devices of a specific deviceType. Returns an empty vector if none are found

MPIMeasurements.getFast — Method

Fast data command mode query. 0 = on, 1 = off

MPIMeasurements.getFirmwareVersion — Method

Returns the firmware version.

MPIMeasurements.getMeasuringChannels — Method

Returns the index of the active channel.

MPIMeasurements.getMode — Method

Returns the mode of the active channel DC = 0 AC = 1

MPIMeasurements.getModelName — Method

Returns the model name of the device.

MPIMeasurements.getPeakRMS — Method

Returns the AC Mode RMS = 0, Peak = 1

MPIMeasurements.getProbeType — Method

Returns the type of the probe 0 = High Sensitivity (HSE) 1 = High Stability (HST) 2 = Ultra-High Sensitivity (UHS)

MPIMeasurements.getRange — Method

Returns the range of the active channel. The range depends on the installed probe. For HSE Probe. More Information in part 3.4 on page 3-7. 0 = highest = ±3T 1 = ±300mT 2 = ±30mT 3 = lowest = ±3mT

MPIMeasurements.getSdoObject — Method

getSdoObject(rob::IgusRobot, sdoObject::SDOObj)

Queries the rob and returns the value of the chosen sdoObject

MPIMeasurements.getSequenceList — Method

getSequenceList(scanner)

Retrieve a list of all sequences available for the scanner.

MPIMeasurements.getSerialDevices — Method

Read out current Serial Ports, returns Array{String,1}

MPIMeasurements.getSerialNumber — Method

Returns the model name of the device.

MPIMeasurements.getTemperature — Method

Returns the current temperature of a channel.

MPIMeasurements.getTemperature — Method

Returns the current temperature of all channel.

MPIMeasurements.getTemperatureAveraging — Method

Returns the averaging count of a given channel.

MPIMeasurements.getUnit — Method

Returns the active used unit G = gauss, T = tesla

MPIMeasurements.getXYZValues — Method

Returns x,y, and z values and applies a coordinate transformation

MPIMeasurements.getXYZValues — Method

Returns x,y, and z values and apply a coordinate transformation

MPIMeasurements.hasDependency — Method

Check whether the device has a dependency of the given type.

MPIMeasurements.identification — Method

Returns the manufacturerID, model number, derial number and firmware revision date

MPIMeasurements.init — Function

init(scanner; ...)
init(scanner, deviceIDs; kwargs...)

(Re-)initializes the given devices of a Scanner according to the current configuration file. This also initializes all dependencies of the given devices. This does not initialize devices that themselves depend on the given devices.

MPIMeasurements.initAxes — Function

Initializes all axes in order Z,Y,X

MPIMeasurements.initiateDevices — Method

initiateDevices(configDir, devicesParams; robust, order)

Initiate devices from the given configuration dictionary.

The device types are referenced by strings matching their device struct name. All device structs are supplied with the device ID and the corresponding device configuration struct.

MPIMeasurements.invertAxesYZ — Method

Inverts the axes for y,z

MPIMeasurements.invertAxisZ — Method

Inverts the axis for z

MPIMeasurements.isAutoRanging — Method

Returns the state of auto ranging for the active channel

MPIMeasurements.isExpectedDependency — Method

Check if a device is an expected dependency of another device.

MPIMeasurements.isFieldComp — Method

Returns the state of the field compensation

MPIMeasurements.isInSleepMode — Method

Returns the sleep mode status of the gaussmeter

MPIMeasurements.isLocked — Method

Returns the state of the frontpanel lock

MPIMeasurements.isOptional — Method

Check whether the device is optional.

MPIMeasurements.isPresent — Method

Check whether the device is present.

MPIMeasurements.isProbeOn — Method

Returns the state of the probe of the active channel

MPIMeasurements.isTempComp — Method

Returns the state of the temperatur compensation

MPIMeasurements.lockOff — Method

Unlocks the frontpanel

MPIMeasurements.lockOn — Method

Locks the frontpanel

MPIMeasurements.moveCenter — Method

Move Bruker Robot to center

MPIMeasurements.movePark — Method

Move Bruker Robot to park

MPIMeasurements.name — Method

Name of the scanner

MPIMeasurements.numChannels — Method

Returns the number of channels the device has.

MPIMeasurements.params — Method

Retrieve the parameters of a device.

MPIMeasurements.query — Method

Send querry to serial device and receive device answer. Returns a String

MPIMeasurements.query — Method

Perform a query to the instrument. Parse result as type T.

MPIMeasurements.query — Method

Perform a query to the instrument. Return String.

MPIMeasurements.query — Method

Perform a query to the iunstrument. Parse result as type T.

MPIMeasurements.query — Method

Perform a query to the instrument. Return String.

MPIMeasurements.queryIsel — Function

queryIsel(sd::SerialDevice,cmd::String)

MPIMeasurements.readIOInput — Method

readIOInput(robot::IselRobot,input::Int64) returns an Bool for the input slot

MPIMeasurements.readIOInput — Method

readIOInput(robot::IselRobot) returns an Array{Bool,1} of the 1-8 input slots

MPIMeasurements.readModbusTelegram — Method

readModbusTelegram(telegram::Vector{UInt8})

Checks the received telegram for error codes and returns the data payload

MPIMeasurements.receive — Method

Read out current content of the output buffer of the serial devive. Returns a String.

MPIMeasurements.scannerBoreSize — Method

Bore size of the scanner.

MPIMeasurements.scannerDatasetStore — Method

Path of the dataset store.

MPIMeasurements.scannerFacility — Method

Facility where the scanner is located.

MPIMeasurements.scannerGradient — Method

Gradient of the scanners selection field.

MPIMeasurements.scannerManufacturer — Method

Manufacturer of the scanner.

MPIMeasurements.scannerName — Method

Name of the scanner

MPIMeasurements.scannerTopology — Method

Topology of the scanner, e.g. FFL or FFP.

MPIMeasurements.selfTest — Method

The gaussmeter reports status based on test done at power up. 0 = no erors found, 1= erros found

MPIMeasurements.sendAndReceiveTelegram — Method

sendAndReceiveTelegram(rob::IgusRobot, telegram::Vector{UInt8})

MPIMeasurements.sendCommand — Method

Send Command sendCommand(sd::BrukerRobot, brukercmd::BrukerCommand)

MPIMeasurements.setAcceleration — Method

Sets Acceleration

MPIMeasurements.setActiveChannel — Method

Sets the active channels. The parameter channel encodes the switched on and off channels. The bits of the ASCII encoded hexadecimal bytes are the channels, with bit 0 representing channel 1 to bit 7 for channel 8.

For example channel="1E" will switch on channels 2, 3, 4, and 5 and switch off the remaining channels 1,6,7, and 8.

MPIMeasurements.setActiveChannel — Method

Sets the active channel to X, Y, Z, V = Vector Magnitude Channel

MPIMeasurements.setAllAutoRanging — Method

Set auto ranging of all channels

MPIMeasurements.setAllMode — Method

Set the mode of all channels DC = 0, AC = 1

MPIMeasurements.setAllRange — Method

Sets the range of all channels

MPIMeasurements.setAutoRanging — Method

Set state of auto ranging for the active channel

MPIMeasurements.setBrake — Method

Sets brake, brake=false no current on brake , brake=true current on brake

MPIMeasurements.setCompleteProbe — Method

Sets complete probe on = 0 or off = 1

MPIMeasurements.setFast — Method

Set fast data command mode. 0 = on, 1 = off

MPIMeasurements.setFieldComp — Method

Set the state of the field compensation. 1 = on, 0 = off

MPIMeasurements.setFree — Method

Sets free, Freifahren axis, wenn Achse über den Referenzpunkt gefahren ist

MPIMeasurements.setFrontPanelLock — Method

sets the state of the front panel lock. Locks all entries except the alarm keys. 1 = on, 0 = off

MPIMeasurements.setMode — Method

Set the mode for the channel DC = 0 AC = 1

MPIMeasurements.setPeakRMS — Method

Set the AC mode for the channel RMS = 0, Peak = 1

MPIMeasurements.setProbe — Method

Sets probe on = 0 or off = 1 on specific channel

MPIMeasurements.setRange — Method

Sets the range of the active channel. The range depends on the installed probe. For HSE Probe. More Information in part 3.4 on page 3-7. 0 = highest = ±3T 1 = ±300mT 2 = ±30mT 3 = lowest = ±3mT

MPIMeasurements.setRefVelocity — Method

Sets the Reference velocities of the axes x,y,z

MPIMeasurements.setSdoObject — Method

setSdoObject(rob::IgusRobot, sdoObject::SDOObj, value::Integer)

Queries the rob and sets a new value for the chosen sdoObject

MPIMeasurements.setSleepMode — Method

Sets the sleep mode state of the gaussmeter. 0 = on, 1 = off

MPIMeasurements.setStandardSettings — Method

Sets the standard settings -highest range -unit to tesla -auto ranging off -complete probe on

MPIMeasurements.setStartStopFreq — Method

Sets StartStopFrequency

MPIMeasurements.setTempComp — Method

Set the state of the temperatur compensation. 1 = on, 0 = off

MPIMeasurements.setTemperatureAveraging — Method

Sets the moving averaging count of a given channel to 0<numAverages<21.

MPIMeasurements.setTxParams — Function

Set the amplitude and phase for all the selected channels.

Note: amplitudes and phases are defined as a dictionary of vectors, since every channel referenced by the dict's key could have a different amount of components.

MPIMeasurements.setUnit — Method

Sets the active unit to 'G' = gauss, 'T' = tesla

MPIMeasurements.setUnitToGauss — Method

Sets the unit of the values to gauss

MPIMeasurements.setUnitToTesla — Method

Sets the unit of the values to tesla

MPIMeasurements.setXRange — Method

Sets the range of the active channel. The range depends on the installed probe. For HSE Probe. More Information in part 3.4 on page 3-7. 0 = highest = ±3T 1 = ±300mT 2 = ±30mT 3 = lowest = ±3mT

Sets the range of the X channel

MPIMeasurements.setYRange — Method

Sets the range of the Y channel

MPIMeasurements.setZRange — Method

Sets the range of the Z channel

MPIMeasurements.setZeroPoint — Method

Sets the zero position for absolute moving at current axes position Z,Y,X

MPIMeasurements.set_delim_read — Method

Set character which terminates query answer.

MPIMeasurements.set_delim_write — Method

Set character which terminates query.

MPIMeasurements.set_timeout_ms — Method

Set maximal time to wait for query answer in ms.

MPIMeasurements.simRefZYX — Method

Simulates Reference Z,Y,X

MPIMeasurements.simulateLangevinInduced — Method

Very, very basic simulation of an MPI signal using the Langevin function.

MPIMeasurements.sleepModeOff — Method

Sets the sleep mode off

MPIMeasurements.sleepModeOn — Method

Sets the sleep mode on

MPIMeasurements.toCurrentMode — Method

Sets the amplifier to current mode.

MPIMeasurements.toHighVoltageMode — Method

Sets the amplifier to high voltage mode.

MPIMeasurements.toLowVoltageMode — Method

Sets the amplifier to low voltage mode.

MPIMeasurements.toVoltageMode — Method

Sets the amplifier to voltage mode.

MPIMeasurements.writeIOOutput — Method

writeIOOutput(robot::IselRobot,output::Array{Bool,1}) output represents 1-8 output slots

Sockets.send — Method

Send command string to serial device.