MPIMeasurements.deviceID — Function

Retrieve the ID of a device.

MPIMeasurements.params — Function

Retrieve the parameters of a device.

MPIMeasurements.isOptional — Function

Check whether the device is optional.

MPIMeasurements.isPresent — Function

Check whether the device is present.

MPIMeasurements.dependencies — Method

Retrieve the dependencies of a device.

MPIMeasurements.dependencies — Method

Retrieve all dependencies of a certain type.

MPIMeasurements.dependency — Method

Retrieve a single dependency of a certain type and error if there are more dependencies.

MPIMeasurements.dependency — Method

Retrieve a single dependency with a certain device ID.

Missing docstring.

Missing docstring for MPIMeasurements.hasdependency(::Device, ::DataType). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPIMeasurements.neededDependencies. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPIMeasurements.optionalDependencies. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPIMeasurements.init(::Device). Check Documenter's build log for details.

Missing docstring.

Missing docstring for MPIMeasurements.close(::Device). Check Documenter's build log for details.